+++ pipelines/generate_test_data.py
@@ -15,7 +15,7 @@ from scipy.stats import expon, norm
 from pipelines.prep_data import process_and_save_loc
 from pipelines.prep_ww_data import clean_nwss_data, preprocess_ww_data
 from pipelines.utils import get_model_data_and_priors_from_dir
-from pyrenew_hew.util import build_pyrenew_model
+from pyrenew_hew.utils import build_pyrenew_model
 
 parser = argparse.ArgumentParser(
     description="Create fit data for disease modeling."
