+++ pipelines/tests/test_build_pyrenew_model.py
@@ -5,7 +5,7 @@ import polars as pl
 import pytest
 
 from pipelines.utils import get_model_data_and_priors_from_dir
-from pyrenew_hew.util import build_pyrenew_model
+from pyrenew_hew.utils import build_pyrenew_model
 
 
 @pytest.fixture
